mlService/requirements.txt
grpcio
grpcio-tools
httpx
chromadb-client
